Reference

Generic

AOCoptimizer.AOCoptimizer — Module

AOCoptimizer is a Julia package for solving QUBO (Quadratic Unconstrained Binary Optimization) and QUMO (Quadratic Unconstrained Binary Optimization) problems.

AOCoptimizer.Direction — Type

Direction

An enum of possible values for the direction of optimization. The optimizer can either minimize (MINIMIZATION) or maximize (MAXIMIZATION) the objective function.

Metrics

AOCoptimizer.hamiltonian — Function

hamiltonian(matrix::AbstractMatrix, x::AbstractVector)
hamiltonian(matrix::AbstractMatrix, field::Union{Nothing,AbstractVector}, x::AbstractVector)

Computes the Hamiltonian of system with interactions expressed in matrix and for a given assignment x. Optionally, a field can be provided. The matrix is expected to be symmetric.

Warning

The matrix is expected to be symmetric. This is not checked. The field is expected to be a vector of the same size as x. This is not checked.

Warning

We expect that the backend storage for matrix and x is the same. This is not checked.

AOCoptimizer.graph_cut_from_hamiltonian — Function

graph_cut_from_hamiltonian(T::Type{<:Real}, graph::AbstractMatrix, hamiltonian::Real)
graph_cut_from_hamiltonian(graph::AbstractMatrix, hamiltonian::Real)
graph_cut_from_hamiltonian(sum_of_matrix_weights::Real, hamiltonian::Real)

Compute the cut of a graph (represented as a matrix graph, or by providing the sum_of_matrix_weights) from its hamiltonian energy.

Tip

If the initial object is a graph, then the sum_of_matrix_weights is the double of the sum of the weights of the graph.

Danger

In addition to assuming that the graph is symmetric (it is not checked), the computation also assumes that the hamiltonian has been computed from the same graph and using a vector assignment of 1 and -1 values.

Runtime system

AOCoptimizer.Environment.local_system_info — Function

local_system_info()::Dict{String,Any}

Capture the local system information and return it as a dictionary.

Multi-threading

AOCoptimizer.CancellationToken — Type

Cancellation tokens can be used to enable the caller of a task to cancel the task execution

AOCoptimizer.RuntimeUtils.RuntimeException — Type

RuntimeException

Exception type for runtime errors.

AOCoptimizer.create_cancellation_token — Function

create_cancellation_token()

Create a cancellation token.

AOCoptimizer.cancel! — Function

cancel!(ctx)

Send a cancellation message using cancellation token ctx.

AOCoptimizer.is_cancelled — Function

is_cancelled(ctx)

Checks whether cancellation token ctx has been cancelled

AOCoptimizer.RuntimeUtils.run_for — Function

run_for(fn, timeout; threads)

Executes function fn for timeout seconds using threads parallel threads. The function fn should accept just one argument, which is a cancellation token; it can return any value. Returns the output of all executions.

QUBO

AOCoptimizer.QUBO.qubo — Type

qubo{T}

A structure representing a Quadratic Unconstrained Binary Optimization (QUBO) problem. The qubo structure contains the following fields:

Sense: The direction of optimization, either MINIMIZATION or MAXIMIZATION.

Terms: A matrix representing the quadratic and linear terms of the QUBO problem.

Assuming that $Q_{i,j} = q.Terms[i ,j]$ is the matrix of quadratic terms for q an instance of qubo{T}, the QUBO problem can be expressed as finding a vector of binary variables $x = (x_1, x_2, \ldots, x_n)$ that minimizes (or, maximizes) the function:

\[f_Q(x) = \sum_{i=1}^{n}\sum_{j=1}^{n} Q_{i,j} x_i x_j\]

Observe that diagonal elements of the matrix $Q$ (q.Terms) represent linear terms. This economizes space and operations without changing the semantics of the problem. The semantics do not change since $w_ii * x_i^2 = w_ii * x_i$ for any $x_i$ in ${0, 1}$ (assuming that $w_ii$ are the diagonal elements of the matrix). However, it is important to adjust the weights of the diagonal elements to express correctly the function $f_Q(x)$. Since $w_{ij}=w_{ji}$, the interaction between the variables $x_i$ and $x_j$ (for $i\ne j$) is added twice in $f_Q(x)$, once in the term $w_{ij} x_i x_j$ and once in $w_{ji} x_j x_i$. On the other hand there is a single term for the diagonal elements $w_{ii} x_i^2 = w_{ii} x_i.$

Warning

Consider doubling the weights of the diagonal elements, if you want to encode an external field!

AOCoptimizer.QUBO.size — Function

size(q::qubo)::Integer

Returns the size of the QUBO problem represented by q.

AOCoptimizer.QUBO.evaluate — Function

evaluate(q::qubo, assignment)

Evaluates the objective function of the QUBO problem q for a given assignment. Returns the value of the objective function.

AOCoptimizer.QUBO._random — Function

_random(rng::AbstractRNG, q::qubo{T})::AbstractVector{T} where {T<:Real}

Generates a random assignment for the QUBO problem q using the provided random number generator rng.

AOCoptimizer.QUBO.greedy_random — Function

greedy_random(rng::Union{AbstractRNG, Integer}, q::qubo{T}, ctx::Union{Nothing,CancellationToken})

Performs a greedy random search on the QUBO problem q using the provided random number generator rng. The search continues until the cancellation token ctx is triggered. Returns a tuple containing the best energy found and the corresponding assignment.

AOCoptimizer.QUBO.increase! — Function

increase!(assignment, terms, indices; ctx)

Tries to increase the objective function of an assignment applied to a set of QUBO terms by flipping assignments one-by-one according to the order in indices.

If cancellation token ctx is present, then process will continue until the entire list of indices gets visited, or cancellation token ctx gets triggered.

Method returns true if progress has been made; false otherwise

AOCoptimizer.QUBO.decrease! — Function

decrease(assignment, terms, indices; ctx)

Tries to decrease the objective function of an assignment applied to a set of QUBO terms by flipping assignments one-by-one according to the order in indices.

If cancellation token ctx is present, then process will continue until the entire list of indices gets visited, or cancellation token ctx gets triggered.

Method returns true if progress has been made; false otherwise

QUMO

AOCoptimizer.QUMO.mixed_ising — Type

mixed_ising{T}

A structure representing a Mixed Ising problem. Such problems are always minimization problems and, unlike traditional Ising problems, they can contain both binary and continuous variables. The continuous variables have a range of [-1, 1], while the binary variables are in {-1, 1}.

The mixed_ising structure contains:

Quadratic: A matrix representing the quadratic terms of the Ising problem.

Field: A vector representing the linear terms of the Ising problem.

Binary: A vector of booleans indicating whether each variable is binary.

Offset: A constant offset added to the objective function.

AOCoptimizer.QUMO.qumo — Type

qumo{T}

A structure representing a Quadratic Unconstrained Mixed Optimization (QUMO). Such problems can be either minimization or maximization problems. They can contain both binary and continuous variables, but the continuous variables are expected to be in the range [-1, 1] and the binary variables are in {0, 1}. The objective function is (1/2)x'Qx + f'x + c, where x is the vector of variables, Q is the quadratic matrix, f is the linear vector (if it exists), and c is a constant offset.

The qumo structure contains the following fields:

Sense: The direction of optimization, either MINIMIZATION or MAXIMIZATION.

Quadratic: A matrix representing the quadratic terms of the QUMO problem.

Field: A vector representing the linear terms of the QUMO problem.

Binary: A vector of booleans indicating whether each variable is binary.

Offset: A constant offset added to the objective function.

It is expected that the Quadratic matrix is symmetric and that the diagonal elements that correspond to binary variables are zero. (This is not checked in the constructor.)

Base.isapprox — Function

isapprox(q1::qumo, q2::qumo; kwargs...)

Checks if two qumo objects are approximately equal, i.e., have the same optimization direction and approximately the same weights in the quadratic and linear terms. They should also have the same binary and continuous variables.

AOCoptimizer.QUMO.number_of_variables — Function

number_of_variables(q::Union{qumo{T}, mixed_ising{T}})::Integer where {T<:Real}

Returns the number of variables in the QUMO or Mixed Ising problem q.

AOCoptimizer.QUMO.convert_mixed_ising_to_positive_qumo — Function

convert_mixed_ising_to_positive_qumo(m::mixed_ising{T})::qumo{T} where T<:Real

Converts a mixed Ising problem, i.e. a problem where binaries are ${-1, 1}$ and continuous in $[-1, 1]$ to positive QUMO, where all variables are expected to be positive (binary in ${0,1}$, continuous in $[0,1]$).

The transformation is achieved by observing that we need to scale all variables $x$ by $(x+1)/2$ to move from $[-1,1]$ to $[0,1]$. Observe, that the optimization objective is $(1/2)x'Qx + f'x + c$.

AOCoptimizer.QUMO.convert_positive_qumo_to_mixed_ising — Function

convert_positive_qumo_to_mixed_ising(q::qumo{T})::mixed_ising{T} where {T<:Real}

Converts a positive QUMO problem, i.e. a problem where all variables are expected to be positive (binary in ${0,1}$, continuous in $[0,1]$) to mixed Ising, where binaries are ${-1, 1}$ and continuous in $[-1, 1]$.

The transformation is achieved by observing that we need to scale all variables $x$ by $2x-1$ to move from $[0,1]$ to $[-1,1]$. Observe, that the optimization objective is $(1/2)x'Qx + f'x + c$.

AOCoptimizer.QUMO.convert_to_qumo — Function

convert_to_qumo(q::mixed_ising{T})::qumo{T} where {T<:Real}

Converts a mixed Ising problem, i.e., a problem where binaries are ${-1, 1}$ and continuous in $[-1, 1]$ to QUMO, where binaries are in ${0, 1}$ and continuous stay at $[-1, 1]$.

AOCoptimizer.QUMO.convert_to_mixed_ising — Function

convert_to_mixed_ising(q::qumo{T})::mixed_ising{T} where {T<:Real}

Converts a QUMO problem to a mixed Ising problem. The transformation is achieved by observing that we need to map all binary variables from {0, 1} to {-1, 1}.

Solver

Energy computations

AOCoptimizer.Solver.calculate_energies! — Function

calculate_energies!(
    energies::AbstractVector{TEval},
    spins::AbstractMatrix{T},
    matrix::AbstractMatrix{TEval},
    external::Union{Nothing,AbstractVector{TEval}},
) where {T<:Real,TEval<:Real}

Calculate the energies for each configuration given the spins, the interaction matrix and an optional external field. The energies vector is modified in place to hold the computed energies.

AOCoptimizer.Solver.calculate_energies — Function

calculate_energies(
    spins::AbstractMatrix{T},
    matrix::AbstractMatrix{TEval},
    external::Union{Nothing,AbstractVector{TEval}},
)::AbstractVector{TEval} where {T<:Real,TEval<:Real}

Calculate the energies for each configuration given the spins, the interaction matrix and an optional external field. The result is returned as a new vector of energies.

AOCoptimizer.Solver.count_min_energy_hits — Function

count_min_energy_hits(observations::TEnergyObservations{<:Real})::AbstractVector{T}

Count the number of experiments that hit the minimum observed energy. In the input matrix, each column corresponds to a configuration and each row corresponds to an experiment. The output is a vector of length equal to the number of columns (configurations) in the input matrix. The entry at index i is the number of experiments that hit the minimum observed energy for configuration i.

Non-linearities and walls

AOCoptimizer.Solver.@make_non_linearity — Macro

@make_non_linearity(name, fn)

Macro to create a non-linearity function that applies the function fn element-wise to an array x.

Where supported, it will generate optimized code for specific backends (e.g., CUDA).

AOCoptimizer.Solver.non_linearity_tanh! — Function

non_linearity_tanh!(x::AbstractArray{T,N}) where {T<:Real,N}

In-place application of the sign function element-wise to the array x.

AOCoptimizer.Solver.non_linearity_sign! — Function

non_linearity_sign!(x::AbstractArray{T,N}) where {T<:Real,N}

In-place application of the sign function element-wise to the array x.

AOCoptimizer.Solver.non_linearity_binary! — Function

non_linearity_binary!(x::AbstractArray{T,N}) where {T<:Real,N}

In-place application of the function x -> (x>0.5)?1.0:0.0 element-wise to the array x.

AOCoptimizer.Solver.@make_wall — Macro

@make_wall(name, lower, upper)

Creates functions that enforce an inelastic wall for a given name, with specified lower and upper bounds.

AOCoptimizer.Solver.enforce_inelastic_wall! — Function

enforce_inelastic_wall!(x::AbstractArray{T,N}, upper::T, lower::T) where {T<:Real,N}
enforce_inelastic_wall!(x::AbstractArray{T,N}, y::AbstractArray{T,N}, upper::T, lower::T) where {T<:Real,N}

Enforces that the elements of x are within the specified lower and upper bounds. If y is provided, it is modified to zero for the corresponding elements for which x exceeds the bounds.

Samplers

AOCoptimizer.Solver.@make_sampler — Macro

@make_sampler(
    name,
    binary_non_linearity,
    walls,
    bias,
    inplace_matrix_vector_multiplication,
    adjust_parameters = nothing,
    per_iteration_callback = nothing
)

This macro generates a sampler function. The generated functions will be named name!. It will use the functions binary_non_linearity and walls to apply the binary non-linearity and the bounds to the variables, respectively. The inplace_matrix_vector_multiplication function will be used to perform the matrix-vector multiplication in-place, which is crucial for performance. It allows to subtract a custom bias from all variables (typically zero, but can be non-zero when simulating physical systems).

The caller can also provide two extensibility points: (A) adjust_parameters, which allows to adjust the gradient and momentum; (B) per_iteration_callback, which allows to run a callback at the end of each iteration.

AOCoptimizer.Solver.sample! — Function

sample!(
    problem::Problem{T, TEval},
    setup::Setup{T},
    workspace::Workspace{T},
    iterations::Integer,
    annealing_delta::AbstractVector{T},
    per_iteration_callback_state::Union{Nothing,TIterationCallbackState}=nothing
) where {T<:Real, TEval<:Real, TIterationCallbackState}

Sampler for mixed Ising problems. Arguments include:

problem contains the input problem. The related matrices will not change. The computation will be performed using the T elementary data-type, but the evaluation of the energies will be performed using the TEval data-type.

setup contains the setup parameters, such as the gradient, momentum, and dt. This will not change during the sampling process.

workspace contains caller-provided scratch buffers that will be used to perform the sampling. This buffers will change during the sampling process, but the caller should not depend on the values in them after the sampling is done.

iterations is the number of iterations to perform.

annealing_delta is a vector of the annealing factor values to use. This vector will change during the sampling process, and at the end it will typically contain zeros.

per_iteration_callback_state is an optional state that will be passed to the per-iteration callback function, if it is provided. This state can be used to collect statistics or to perform other actions at the end of each iteration.

AOCoptimizer.Solver.sample_qumo! — Function

sample_qumo!(
    problem::Problem{T, TEval},
    setup::Setup{T},
    workspace::Workspace{T},
    iterations::Integer,
    annealing_delta::AbstractVector{T},
    per_iteration_callback_state::Union{Nothing,TIterationCallbackState}=nothing
) where {T<:Real, TEval<:Real, TIterationCallbackState}

Sampler for mixed QUMO problems, where the binaries are {0, 1} and the continuous in [-1, 1]. Arguments include:

problem contains the input problem. The related matrices will not change. The computation will be performed using the T elementary data-type, but the evaluation of the energies will be performed using the TEval data-type.

setup contains the setup parameters, such as the gradient, momentum, and dt. This will not change during the sampling process.

workspace contains caller-provided scratch buffers that will be used to perform the sampling. This buffers will change during the sampling process, but the caller should not depend on the values in them after the sampling is done.

iterations is the number of iterations to perform.

annealing_delta is a vector of the annealing factor values to use. This vector will change during the sampling process, and at the end it will typically contain zeros.

per_iteration_callback_state is an optional state that will be passed to the per-iteration callback function, if it is provided. This state can be used to collect statistics or to perform other actions at the end of each iteration.

AOCoptimizer.Solver.sample_binary! — Function

sample_binary!(
    problem::Problem{T, TEval},
    setup::Setup{T},
    workspace::Workspace{T},
    iterations::Integer,
    annealing_delta::AbstractVector{T},
    per_iteration_callback_state::Union{Nothing,TIterationCallbackState}=nothing
) where {T<:Real, TEval<:Real, TIterationCallbackState}

Sampler for mixed positive QUMO problems, where the binaries are {0, 1} and the continuous in [0, 1]. Arguments include:

problem contains the input problem. The related matrices will not change. The computation will be performed using the T elementary data-type, but the evaluation of the energies will be performed using the TEval data-type.

setup contains the setup parameters, such as the gradient, momentum, and dt. This will not change during the sampling process.

workspace contains caller-provided scratch buffers that will be used to perform the sampling. This buffers will change during the sampling process, but the caller should not depend on the values in them after the sampling is done.

iterations is the number of iterations to perform.

annealing_delta is a vector of the annealing factor values to use. This vector will change during the sampling process, and at the end it will typically contain zeros.

per_iteration_callback_state is an optional state that will be passed to the per-iteration callback function, if it is provided. This state can be used to collect statistics or to perform other actions at the end of each iteration.

AOCoptimizer.Solver.SamplerTracer.Periodic — Type

Periodic{TM<:AbstractArray{<:Real,3}}

A tracer that collects the spins at regular intervals.

AOCoptimizer.Solver.SamplerTracer.SamplerWithPlan — Type

SamplerWithPlan{TM<:AbstractArray{<:Real,3}}

A tracer that collects the spins at specific iterations, explicitly provided by the user. These sampling points are typically related to the number of iterations that we expect the sampler to run for.

AOCoptimizer.Solver.sample_with_tracer! — Function

sample_with_tracer!(
    problem::Problem{T, TEval},
    setup::Setup{T},
    workspace::Workspace{T},
    iterations::Integer,
    annealing_delta::AbstractVector{T},
    per_iteration_callback_state::Union{Nothing,TIterationCallbackState}=nothing
) where {T<:Real, TEval<:AbstractMatrix{<:T}, TIterationCallbackState}

Exploration

AOCoptimizer.Solver.@make_exploration — Macro

make_exploration(name, sampler, assignment_collector = Collector.BestAssignmentCollector())

Macro to create an exploration function with the given name and sampler. The assignment_collector is used to specify which samples to collect during the exploration. By default, it uses Collector.BestAssignmentCollector(), which collects the best assignment found during the exploration.

AOCoptimizer.Solver.explore — Function

explore(
    problem::Problem{T,TEval},
    initial_setup::Setup{T},
    batch_size::Integer,
    ctx::CancellationToken,
    iterations::Integer,
    repetitions::Integer,
    rng::AbstractRNG
)
explore(
    problem::Problem{T,TEval},
    initial_setup::Setup{T},
    batch_size::Integer,
    ctx::CancellationToken,
    iterations::TIterationNumberChooser,
    repetitions::Integer,
    rng::AbstractRNG
)

Run the exploration for a mixing-Ising problem with the given problem and initial_setup. The exploration will run for the specified number of iterations and repetitions. The batch_size specifies how many samples to process in parallel. The ctx is a cancellation token that can be used to terminate the exploration (e.g., when the allotted time is over). The rng is the random number generator to use for sampling.

AOCoptimizer.Solver.explore_qumo — Function

explore_qumo(
    problem::Problem{T,TEval},
    initial_setup::Setup{T},
    batch_size::Integer,
    ctx::CancellationToken,
    iterations::Integer,
    repetitions::Integer,
    rng::AbstractRNG
)
explore_qumo(
    problem::Problem{T,TEval},
    initial_setup::Setup{T},
    batch_size::Integer,
    ctx::CancellationToken,
    iterations::TIterationNumberChooser,
    repetitions::Integer,
    rng::AbstractRNG
)

Run the exploration for a QUMO problem with the given problem and initial_setup. The exploration will run for the specified number of iterations and repetitions. The batch_size specifies how many samples to process in parallel. The ctx is a cancellation token that can be used to terminate the exploration (e.g., when the allotted time is over). The rng is the random number generator to use for sampling.

AOCoptimizer.Solver.explore_binary — Function

explore_binary(
    problem::Problem{T,TEval},
    initial_setup::Setup{T},
    batch_size::Integer,
    ctx::CancellationToken,
    iterations::Integer,
    repetitions::Integer,
    rng::AbstractRNG
)
explore_binary(
    problem::Problem{T,TEval},
    initial_setup::Setup{T},
    batch_size::Integer,
    ctx::CancellationToken,
    iterations::TIterationNumberChooser,
    repetitions::Integer,
    rng::AbstractRNG
)

Run the exploration for a positive-QUMO problem with the given problem and initial_setup. The exploration will run for the specified number of iterations and repetitions. The batch_size specifies how many samples to process in parallel. The ctx is a cancellation token that can be used to terminate the exploration (e.g., when the allotted time is over). The rng is the random number generator to use for sampling.

AOCoptimizer.Solver.ExplorationResult — Type

ExplorationResult{T <: Real, TV <: AbstractVector, TM <: AbstractMatrix}

The result of an exploration phase. It supports the opaque field CollectorAdditionalInfo, which may be used by the collector to store additional information.

AOCoptimizer.Solver.Collector.BestFound — Type

BestFound{T<:Real, TV<:AbstractVector}

A structure to hold the best found solution and its objective value. All collectors return this structure (when calling the retrieve method on them).

Observe that the elementary type of the assignment vector can be different from the type of the objective value. For example, the objective may be computed at higher precision (e.g., Float32) than the assignment vector (e.g., Float16).

AOCoptimizer.Solver.Collector.BestSolutionState — Type

BestSolutionState{T<:Real, TV<:AbstractVector}

A state structure to hold the best solution found so far. This structure can be updated with new solutions, and it will keep the best one found so far.

Observe that the elementary type of the assignment vector can be different from the type of the objective value. For example, the objective may be computed at higher precision (e.g., Float32) than the assignment vector (e.g., Float16).

Solver initialization

AOCoptimizer.Solver.AbstractProblem — Type

AbstractProblem{T<:Real}

Abstract type for specifying the problem to be solved.

AOCoptimizer.Solver.Problem — Type

Problem{T<:Real} <: AbstractProblem{T}

Data structure for specifying the problem to be solved. This is the case where the problem has an external field. The problem is specified by the interaction matrix and the external field. The problem can contain both binary and continuous variables. It is assumed that the binary variables are the first indices in the matrix and in the field. It is also required that the diagonal of the interaction matrix for the binary variables is zero. For the continuous, the diagonal can be non-zero, as it represents quadratic terms.

AOCoptimizer.Solver.make_annealing_delta — Function

make_annealing_delta(annealing::AbstractVector, iterations::Integer)
make_annealing_delta(setup::Setup, iterations::Integer)

Returns a vector of the values to be used for the annealing parameter.

AOCoptimizer.Solver.Setup — Type

Setup{T<:Real}

This structure holds the configuration parameters that will be used in the solver. The Annealing, Gradient, and Momentum vectors must all have the same length; for each index i, the i-th element of each vector corresponds to the parameters for the i-th experiment. The dt parameter is the time step used in the solver.

AOCoptimizer.Solver.make_empty_setup — Function

make_empty_setup(setup::Setup{T}, capacity::Integer) where {T<:Real}

Creates an empty Setup instance with the same structure as setup, but with capacity elements in each vector. The vectors are filled with zeros.

AOCoptimizer.Solver.make_setup — Function

make_setup(
    annealing::TV,
    gradient::TV,
    momentum::TV,
    dt::T,
    repetitions::Integer,
) where {T<:Real,TV<:AbstractVector{T}}
function make_setup(
    annealing::TV,
    gradient::TV,
    momentum::T,
    dt::T,
    repetitions::Integer,
) where {T<:Real,TV<:AbstractVector{T}}
function make_setup(
    T::Type{<:Real},
    annealing::AbstractVector{<:Real},
    gradient::AbstractVector{<:Real},
    momentum::AbstractVector{<:Real},
    dt::Real,
    repetitions::Integer,
)
function make_setup(
    T::Type{<:Real},
    annealing::AbstractVector{<:Real},
    gradient::AbstractVector{<:Real},
    momentum::Real,
    dt::Real,
    repetitions::Integer,
)

Creates a Setup instance with the specified annealing, gradient, and momentum vectors, repeated repetitions times. The dt parameter is the time step used in the solver. If T is specified, it converts the vectors to that type.

AOCoptimizer.Solver.Workspace — Type

Workspace{<:Real}

A workspace contains all the temporaries that the sampler needs. The temporaries are allocated once and can be reused for multiple samples. Use the make_workspace function to create a workspace, and the initialize_workspace to initialize it between re-uses.

All the matrices below have dimensions equal to the number of variables x number of experiments. The annealing vector has length equal to the number of experiments.

AOCoptimizer.Solver.make_workspace — Function

make_workspace(problem::Problem, setup::Setup, number_of_samples::Integer)::Workspace
make_workspace(problem::Problem, setup::Setup)::Workspace

Creates a workspace, i.e., a collection of temporary buffers, to be used by the sampler.

AOCoptimizer.Solver.initialize_workspace! — Function

initialize_workspace!(
    rng::AbstractRNG,
    workspace::Workspace{T},
    annealing::AbstractVector{T},
    sampler::TSampler
) where {T<:Real}

Initializes a workspace by assigning random values to the initial spins. Caller must also pass the initial values for the annealing terms (typically, from the Setup{T} structure), and the sampler for the initial values to use.

AOCoptimizer.Solver.ConfigurationSpace — Type

ConfigurationSpace{T<:Real}

Defines a configuration space, i.e., the set of coefficients for the parameters of the algorithm we are interested in (parameters are annealing, gradient, and momentum coefficients).

AOCoptimizer.Solver.sample_configuration_space — Function

sample_configuration_space(T, number_of_configurations;
                           annealing::ClosedInterval,
                           gradient::ClosedInterval,
                           momentum::ClosedInterval)

Generates a set of number_of_configurations points that sample the space of valid coefficients for the parameters used by the sampler. The valid values of coefficients are determined by the intervals annealing, gradient and momentum.

Main solver methods and analysis of results

AOCoptimizer.Solver.TEnergyObservations — Type

TEnergyObservations{T<:Number}

Type alias for a matrix of energy observations for various configurations. Each column corresponds to a configuration and each row corresponds to an experiment.

AOCoptimizer.Solver.@make_solver — Macro

@make_solver(name, exploration)

Macro to create a solver function that runs the AOC optimizer. The solver will be named name and will use the exploration function.

AOCoptimizer.Solver.solve — Function

solve(
    T::DataType,
    interactions::AbstractMatrix{TInput},
    field::Union{Nothing,AbstractVector{TInput}},
    binary::Integer,
    timeout::Second;
    rng::AbstractRNG = Random.GLOBAL_RNG,
    engine::Engine = get_current_engine(),
    annealing = ClosedInterval(0.01, 1.0),
    gradient = ClosedInterval(0.01, 1.0),
    momentum = ClosedInterval(0.95, 0.99),
    deep_search_iterations = ClosedInterval(500, 20000),
    dt::Real = 0.5,
    phase_1_fraction::Real = 0.1,
    phase_2_fraction::Real = 0.2,
)

solve(
    T::DataType,
    interactions::AbstractMatrix{TInput},
    timeout::Second;
    rng::AbstractRNG = Random.GLOBAL_RNG,
    engine::Engine = get_current_engine(),
    annealing = ClosedInterval(0.01, 1.0),
    gradient = ClosedInterval(0.01, 1.0),
    momentum = ClosedInterval(0.95, 0.99),
    deep_search_iterations = ClosedInterval(500, 20000),
    dt::Real = 0.5,
    phase_1_fraction::Real = 0.1,
    phase_2_fraction::Real = 0.2,
)

Invokes the AOC solver on a mixed-Ising problem, described by the interactions matrix and the field vector (if provided). The computation is performed with the T primitive type (e.g., Float32, Float64, Float16). The first binary variables in the interactions matrix and in the field vector are treated as binary variables. The first binary diagonal elements of the interactions matrix must be zero. The interactions matrix must be symmetric.

The computation will be performed in the backend backend, which can be CPU, CUDABackend, etc.

The solver will run for timeout seconds. That time is divided into three phases:

Phase 1: Exploration of the configuration space with a limited number of iterations; this phase will take phase_1_fraction of the total time.

Phase 2: Exploration of the configuration space with a larger number of iterations; this phase will take phase_2_fraction of the total time.

Deep search: A deep search phase, which will run for the remaining time. The number of iterations in this phases is chosen with the deep_search_iterations parameter. The number of iterations may be adjusted based on the capabilities of the backend.

The configuration space is sampled from the intervals specified in the annealing, gradient, and momentum parameters.

The dt parameter specifies the time step for the simulation.

AOCoptimizer.Solver.solve_qumo — Function

solve_qumo(
    T::DataType,
    interactions::AbstractMatrix{TInput},
    field::Union{Nothing,AbstractVector{TInput}},
    binary::Integer,
    timeout::Second;
    rng::AbstractRNG = Random.GLOBAL_RNG,
    engine::Engine = get_current_engine(),
    annealing = ClosedInterval(0.01, 1.0),
    gradient = ClosedInterval(0.01, 1.0),
    momentum = ClosedInterval(0.95, 0.99),
    deep_search_iterations = ClosedInterval(500, 20000),
    dt::Real = 0.5,
    phase_1_fraction::Real = 0.1,
    phase_2_fraction::Real = 0.2,
)

solve_qumo(
    T::DataType,
    interactions::AbstractMatrix{TInput},
    timeout::Second;
    rng::AbstractRNG = Random.GLOBAL_RNG,
    engine::Engine = get_current_engine(),
    annealing = ClosedInterval(0.01, 1.0),
    gradient = ClosedInterval(0.01, 1.0),
    momentum = ClosedInterval(0.95, 0.99),
    deep_search_iterations = ClosedInterval(500, 20000),
    dt::Real = 0.5,
    phase_1_fraction::Real = 0.1,
    phase_2_fraction::Real = 0.2,
)

Invokes the AOC solver on a mixed-Ising problem, described by the interactions matrix and the field vector (if provided). The computation is performed with the T primitive type (e.g., Float32, Float64, Float16). The first binary variables in the interactions matrix and in the field vector are treated as binary variables. The first binary diagonal elements of the interactions matrix must be zero. The interactions matrix must be symmetric.

The computation will be performed in the backend backend, which can be CPU, CUDABackend, etc.

The solver will run for timeout seconds. That time is divided into three phases:

Phase 1: Exploration of the configuration space with a limited number of iterations; this phase will take phase_1_fraction of the total time.

Phase 2: Exploration of the configuration space with a larger number of iterations; this phase will take phase_2_fraction of the total time.

Deep search: A deep search phase, which will run for the remaining time. The number of iterations in this phases is chosen with the deep_search_iterations parameter. The number of iterations may be adjusted based on the capabilities of the backend.

The configuration space is sampled from the intervals specified in the annealing, gradient, and momentum parameters.

The dt parameter specifies the time step for the simulation.

AOCoptimizer.Solver.solve_binary — Function

solve_binary(
    T::DataType,
    interactions::AbstractMatrix{TInput},
    field::Union{Nothing,AbstractVector{TInput}},
    binary::Integer,
    timeout::Second;
    rng::AbstractRNG = Random.GLOBAL_RNG,
    engine::Engine = get_current_engine(),
    annealing = ClosedInterval(0.01, 1.0),
    gradient = ClosedInterval(0.01, 1.0),
    momentum = ClosedInterval(0.95, 0.99),
    deep_search_iterations = ClosedInterval(500, 20000),
    dt::Real = 0.5,
    phase_1_fraction::Real = 0.1,
    phase_2_fraction::Real = 0.2,
)

solve_binary(
    T::DataType,
    interactions::AbstractMatrix{TInput},
    timeout::Second;
    rng::AbstractRNG = Random.GLOBAL_RNG,
    engine::Engine = get_current_engine(),
    annealing = ClosedInterval(0.01, 1.0),
    gradient = ClosedInterval(0.01, 1.0),
    momentum = ClosedInterval(0.95, 0.99),
    deep_search_iterations = ClosedInterval(500, 20000),
    dt::Real = 0.5,
    phase_1_fraction::Real = 0.1,
    phase_2_fraction::Real = 0.2,
)

Invokes the AOC solver on a problem where the binary variables are either 0 or 1, and the continuous in the range [0, 1]. The problem is described by the interactions matrix and the field vector (if provided). The computation is performed with the T primitive type (e.g., Float32, Float64, Float16). The first binary variables in the interactions matrix and in the field vector are treated as binary variables. The first binary diagonal elements of the interactions matrix must be zero. The interactions matrix must be symmetric.

The computation will be performed in the backend backend, which can be CPU, CUDABackend, etc.

The solver will run for timeout seconds. That time is divided into three phases:

Phase 1: Exploration of the configuration space with a limited number of iterations; this phase will take phase_1_fraction of the total time.

Phase 2: Exploration of the configuration space with a larger number of iterations; this phase will take phase_2_fraction of the total time.

Deep search: A deep search phase, which will run for the remaining time. The number of iterations in this phases is chosen with the deep_search_iterations parameter. The number of iterations may be adjusted based on the capabilities of the backend.

The configuration space is sampled from the intervals specified in the annealing, gradient, and momentum parameters.

The dt parameter specifies the time step for the simulation.

AOCoptimizer.Solver.get_solver_results_summary — Function

get_solver_results_summary(solver_results)

Computes the relevant statistics for the solver results

AOCoptimizer.Solver.find_best — Function

find_best(results)

Find the best result from the results of a run of solve. It returns the best Hamiltonian and the corresponding assignment in the variables.

AOCoptimizer.Solver.search_for_best_configuration — Function

search_for_best_configuration(results::ExplorationResult)
search_for_best_configuration(results::PhaseStatistics)
search_for_best_configuration(results::TRuntimeInfo)

Searches for the best objective value in the results returned by the solver. It returns the best objective value, as well as other statistics about the solution.

CAUTION: Given the different possible inputs, there is variability in the type of the output. Do not take a strong dependency on the output of this function.

AOCoptimizer.Solver.extract_runtime_information — Function

extract_runtime_information(results)

Extracts only runtime information from the results of the solver (ignoring the solution found, objectives, etc.)

Internal

The following internal methods are not intended for direct use.

AOCoptimizer.Solver._widen_for_eval — Function

_widen_for_eval(T::DataType)

Helper function to determine the data type to use for computing the energies. The main challenge is that evaluating energies using Float16 or BFloat16 leads to numerical inaccuracies and incorrect results. Hence, for picking the best configurations, we use a wider data type.

AOCoptimizer.Solver._find_and_display_negative_values — Function

_find_and_display_negative_values(x::AbstractVector{T}; max_to_print=5) where {T<:Real}

Internal method used for debugging purposes. It finds negative values in the vector x.

AOCoptimizer.Solver._get_time_to_solution — Function

_get_time_to_solution(success_rate, time)

Computes the time to achieve the best seen solution with 99% confidence

AOCoptimizer.Solver._get_num_operations_to_solution — Function

_get_num_operations_to_solution(success_rate, operations_per_deep_search)

Computes the number of matrix-vector multiplications (i.e., basic operations) to perform to achieve the best seen solution with 99% confidence.

AOCoptimizer.Solver._are_all_diags_zero — Function

_are_all_diags_zero(m::AbstractMatrix{T}, binary::Integer) where {T<:Real}

Check if all the diagonal elements of the matrix m are zero. For valid (mixed-)Ising and QUMO/QUBO problems, we require that all diagonal elements that correspond to binary variables ({-1, 1} spins for Ising) are zero; those terms, if necessary, should be folded into the external field (for QUMO/QUBO) or the constant term (for Ising). If binary is provided, it specifies the number of binary variables, and only the first binary diagonal elements are checked.

AOCoptimizer.Solver._MAX_NUMBER_OF_CPU_THREADS — Function

_MAX_NUMBER_OF_CPU_THREADS()::Integer

Returns the number of CPU threads to use when running the solver in the CPU.

AOCoptimizer.Solver._dispose — Function

_dispose(x::Any)

This function serves as a marker to indicate that the object x is not used anymore and can be disposed of.

_dispose(workspace::Workspace)

This function serves mostly as a marker to indicate that the workspace is not used anymore and can be disposed of.

AOCoptimizer.Solver._optimal_batch_size — Function

_optimal_batch_size(backend::Backend, problem::Problem)::Integer
_optimal_batch_size(problem::Problem)::Integer

Internal method to compute the optimal batch size when solving a problem.

AOCoptimizer.Solver.Collector._default_best_assignment_collector — Constant

_default_best_assignment_collector

Internal variable to hold an object of the BestAssignmentCollector type.