AOCoptimizer.FileFormats

AOCoptimizer.FileFormats.FileNotFoundException — Type

FileNotFoundException(filename::AbstractString, message::AbstractString)

Exception thrown when a file does not exist.

AOCoptimizer.FileFormats.GraphIOException — Type

Exception thrown when encountering errors in parsing files

Simple graph file format

AOCoptimizer.FileFormats.read_graph_matrix — Function

read_graph_matrix(io)

Read an undirected graph from an I/O stream.

The format of the file should be:

number_of_vertices number_of_edges
endpoint_A_of_edge_1 endpoint_B_of_edge_1 weight_of_edge_1
endpoint_A_of_edge_2 endpoint_B_of_edge_2 weight_of_edge_2
....

read_graph_matrix(filename)

Read an undirected graph from a file with name filename.

AOCoptimizer.FileFormats.read_directed_graph_matrix — Function

read_directed_graph_matrix(io)

Read a directed graph from an I/O stream.

The format of the file should be:

number_of_vertices number_of_edges
endpoint_A_of_edge_1 endpoint_B_of_edge_1 weight_of_edge_1
endpoint_A_of_edge_2 endpoint_B_of_edge_2 weight_of_edge_2
....

read_directed_graph_matrix(filename)

Read an undirected graph from a file with name filename.

QIO file format

AOCoptimizer.FileFormats.QIO.QIOProblem — Type

QIOProblem

Abstract type for QIO problems.

AOCoptimizer.FileFormats.QIO.Ising — Type

Ising{T<:Real} <: QIOProblem

Ising problem type. Ising problems are represented with a matrix of coefficients and a vector of fields.

AOCoptimizer.FileFormats.QIO.QUBO — Type

QUBO{T<:Real} <: QIOProblem

QUBO problem type. QUBO problems are represented as a matrix of coefficients. These instances do not contain linear terms.

AOCoptimizer.FileFormats.QIO.Metadata — Type

Metadata{T<:Real}

Metadata for QIO problems. Metadata include the following fields:

Objective: The objective value of the problem; this is the best known minimum or maximum value of the objective function.

Instance: The name of the instance; this can be used to identify the problem.

Sense: The optimization direction (MINIMIZATION or MAXIMIZATION).

NumberOfTerms: The number of terms in the problem.

NumberOfInteractions: The number of interactions in the problem.

MaxDegree: The maximum degree of the problem.

AOCoptimizer.FileFormats.QIO.QIOException — Type

QIOException(msg::AbstractString)

Exception thrown when encountering errors in parsing files

AOCoptimizer.FileFormats.QIO.read_qio — Function

read_qio(::Type{T}, filename::AbstractString)::QIOProblem where {T<:Real}
read_qio(filename::AbstractString)::QIOProblem
read_qio(::Type{T}, io::IO)::QIOProblem where {T<:Real}
read_qio(io::IO)::QIOProblem

Read a QIO problem from a file with name filename or from an IO stream io. Optional type parameter T specifies the elementary type used in the the problem (e.g., Float64 or Float32).

AOCoptimizer.FileFormats.QIO.is_qubo — Function

is_qubo(::QIOProblem) -> Bool

Check if the problem is a QUBO problem.

AOCoptimizer.FileFormats.QIO.is_ising — Function

is_ising(::QIOProblem) -> Bool

Check if the problem is an Ising problem.