AOCoptimizer.Algorithms

Enhanced Random

AOCoptimizer.Algorithms.EnhancedRandom.search — Function

search(rng::AbstractRNG, matrix::AbstractMatrix{T}, ctx::CancellationToken)
search(seed::Integer, matrix::AbstractMatrix{T}, ctx::CancellationToken)
search(rng::AbstractRNG, matrix::AbstractMatrix{T}, field::AbstractVector{T}, ctx::CancellationToken)
search(seed::Integer, matrix::AbstractMatrix{T}, field::AbstractVector{T}, ctx::CancellationToken)

Implement a simple heuristic that tries to solve Ising problems, by starting from a number of random solutions and sequentially trying to improve them using single flips.

AOCoptimizer.Algorithms.EnhancedRandom.search_qumo — Function

search_qumo(rng::AbstractRNG, quadratic::QuadraticMatrix{T}, linear::AbstractVector{T}, ctx::CancellationToken)
search_qumo(seed::Integer, quadratic::QuadraticMatrix{T}, linear::AbstractVector{T}, ctx::CancellationToken)
search_qumo(rng::AbstractRNG, quadratic::QuadraticMatrix{T}, ctx::CancellationToken)
search_qumo(seed::Integer, quadratic::QuadraticMatrix{T}, ctx::CancellationToken)

Implement a simple heuristic that tries to solve QUMO problems, by starting from a number of random solutions and sequentially trying to improve them using single flips.