Reference

Generic

AOCoptimizer.AOCoptimizer — Module

AOCoptimizer is a Julia package for solving QUBO (Quadratic Unconstrained Binary Optimization) and QUMO (Quadratic Unconstrained Binary Optimization) problems.

AOCoptimizer.Direction — Type

Direction

An enum of possible values for the direction of optimization. The optimizer can either minimize (MINIMIZATION) or maximize (MAXIMIZATION) the objective function.

Metrics

AOCoptimizer.hamiltonian — Function

hamiltonian(matrix::AbstractMatrix, x::AbstractVector)
hamiltonian(matrix::AbstractMatrix, field::Union{Nothing,AbstractVector}, x::AbstractVector)

Computes the Hamiltonian of system with interactions expressed in matrix and for a given assignment x. Optionally, a field can be provided. The matrix is expected to be symmetric.

Warning

The matrix is expected to be symmetric. This is not checked. The field is expected to be a vector of the same size as x. This is not checked.

Warning

We expect that the backend storage for matrix and x is the same. This is not checked.

AOCoptimizer.graph_cut_from_hamiltonian — Function

graph_cut_from_hamiltonian(T::Type{<:Real}, graph::AbstractMatrix, hamiltonian::Real)
graph_cut_from_hamiltonian(graph::AbstractMatrix, hamiltonian::Real)
graph_cut_from_hamiltonian(sum_of_matrix_weights::Real, hamiltonian::Real)

Compute the cut of a graph (represented as a matrix graph, or by providing the sum_of_matrix_weights) from its hamiltonian energy.

Tip

If the initial object is a graph, then the sum_of_matrix_weights is the double of the sum of the weights of the graph.

Danger

In addition to assuming that the graph is symmetric (it is not checked), the computation also assumes that the hamiltonian has been computed from the same graph and using a vector assignment of 1 and -1 values.

Multi-threading

AOCoptimizer.CancellationToken — Type

Cancellation tokens can be used to enable the caller of a task to cancel the task execution

AOCoptimizer.RuntimeUtils.RuntimeException — Type

RuntimeException

Exception type for runtime errors.

AOCoptimizer.create_cancellation_token — Function

create_cancellation_token()

Create a cancellation token.

AOCoptimizer.cancel! — Function

cancel!(ctx)

Send a cancellation message using cancellation token ctx.

AOCoptimizer.is_cancelled — Function

is_cancelled(ctx)

Checks whether cancellation token ctx has been cancelled

AOCoptimizer.RuntimeUtils.run_for — Function

run_for(fn, timeout; threads)

Executes function fn for timeout seconds using threads parallel threads. The function fn should accept just one argument, which is a cancellation token; it can return any value. Returns the output of all executions.

QUBO

AOCoptimizer.QUBO.qubo — Type

qubo{T}

A structure representing a Quadratic Unconstrained Binary Optimization (QUBO) problem. The qubo structure contains the following fields:

Sense: The direction of optimization, either MINIMIZATION or MAXIMIZATION.

Terms: A matrix representing the quadratic and linear terms of the QUBO problem.

Assuming that $Q_{i,j} = q.Terms[i ,j]$ is the matrix of quadratic terms for q an instance of qubo{T}, the QUBO problem can be expressed as finding a vector of binary variables $x = (x_1, x_2, \ldots, x_n)$ that minimizes (or, maximizes) the function:

\[f_Q(x) = \sum_{i=1}^{n}\sum_{j=1}^{n} Q_{i,j} x_i x_j\]

Observe that diagonal elements of the matrix $Q$ (q.Terms) represent linear terms. This economizes space and operations without changing the semantics of the problem. The semantics do not change since $w_ii * x_i^2 = w_ii * x_i$ for any $x_i$ in ${0, 1}$ (assuming that $w_ii$ are the diagonal elements of the matrix). However, it is important to adjust the weights of the diagonal elements to express correctly the function $f_Q(x)$. Since $w_{ij}=w_{ji}$, the interaction between the variables $x_i$ and $x_j$ (for $i\ne j$) is added twice in $f_Q(x)$, once in the term $w_{ij} x_i x_j$ and once in $w_{ji} x_j x_i$. On the other hand there is a single term for the diagonal elements $w_{ii} x_i^2 = w_{ii} x_i.$

Warning

Consider doubling the weights of the diagonal elements, if you want to encode an external field!

AOCoptimizer.QUBO.size — Function

size(q::qubo)::Integer

Returns the size of the QUBO problem represented by q.

AOCoptimizer.QUBO.evaluate — Function

evaluate(q::qubo, assignment)

Evaluates the objective function of the QUBO problem q for a given assignment. Returns the value of the objective function.

AOCoptimizer.QUBO._random — Function

_random(rng::AbstractRNG, q::qubo{T})::AbstractVector{T} where {T<:Real}

Generates a random assignment for the QUBO problem q using the provided random number generator rng.

AOCoptimizer.QUBO.greedy_random — Function

greedy_random(rng::Union{AbstractRNG, Integer}, q::qubo{T}, ctx::Union{Nothing,CancellationToken})

Performs a greedy random search on the QUBO problem q using the provided random number generator rng. The search continues until the cancellation token ctx is triggered. Returns a tuple containing the best energy found and the corresponding assignment.

AOCoptimizer.QUBO.increase! — Function

increase!(assignment, terms, indices; ctx)

Tries to increase the objective function of an assignment applied to a set of QUBO terms by flipping assignments one-by-one according to the order in indices.

If cancellation token ctx is present, then process will continue until the entire list of indices gets visited, or cancellation token ctx gets triggered.

Method returns true if progress has been made; false otherwise

AOCoptimizer.QUBO.decrease! — Function

decrease(assignment, terms, indices; ctx)

Tries to decrease the objective function of an assignment applied to a set of QUBO terms by flipping assignments one-by-one according to the order in indices.

If cancellation token ctx is present, then process will continue until the entire list of indices gets visited, or cancellation token ctx gets triggered.

Method returns true if progress has been made; false otherwise

QUMO

AOCoptimizer.QUMO.mixed_ising — Type

mixed_ising{T}

A structure representing a Mixed Ising problem. Such problems are always minimization problems and, unlike traditional Ising problems, they can contain both binary and continuous variables. The continuous variables have a range of [-1, 1], while the binary variables are in {-1, 1}.

The mixed_ising structure contains:

Quadratic: A matrix representing the quadratic terms of the Ising problem.

Field: A vector representing the linear terms of the Ising problem.

Binary: A vector of booleans indicating whether each variable is binary.

Offset: A constant offset added to the objective function.

AOCoptimizer.QUMO.qumo — Type

qumo{T}

A structure representing a Quadratic Unconstrained Mixed Optimization (QUMO). Such problems can be either minimization or maximization problems. They can contain both binary and continuous variables, but the continuous variables are expected to be in the range [-1, 1] and the binary variables are in {0, 1}. The objective function is (1/2)x'Qx + f'x + c, where x is the vector of variables, Q is the quadratic matrix, f is the linear vector (if it exists), and c is a constant offset.

The qumo structure contains the following fields:

Sense: The direction of optimization, either MINIMIZATION or MAXIMIZATION.

Quadratic: A matrix representing the quadratic terms of the QUMO problem.

Field: A vector representing the linear terms of the QUMO problem.

Binary: A vector of booleans indicating whether each variable is binary.

Offset: A constant offset added to the objective function.

It is expected that the Quadratic matrix is symmetric and that the diagonal elements that correspond to binary variables are zero. (This is not checked in the constructor.)

Base.isapprox — Function

isapprox(q1::qumo, q2::qumo; kwargs...)

Checks if two qumo objects are approximately equal, i.e., have the same optimization direction and approximately the same weights in the quadratic and linear terms. They should also have the same binary and continuous variables.

AOCoptimizer.QUMO.number_of_variables — Function

number_of_variables(q::Union{qumo{T}, mixed_ising{T}})::Integer where {T<:Real}

Returns the number of variables in the QUMO or Mixed Ising problem q.

AOCoptimizer.QUMO.convert_mixed_ising_to_positive_qumo — Function

convert_mixed_ising_to_positive_qumo(m::mixed_ising{T})::qumo{T} where T<:Real

Converts a mixed Ising problem, i.e. a problem where binaries are ${-1, 1}$ and continuous in $[-1, 1]$ to positive QUMO, where all variables are expected to be positive (binary in ${0,1}$, continuous in $[0,1]$).

The transformation is achieved by observing that we need to scale all variables $x$ by $(x+1)/2$ to move from $[-1,1]$ to $[0,1]$. Observe, that the optimization objective is $(1/2)x'Qx + f'x + c$.

AOCoptimizer.QUMO.convert_positive_qumo_to_mixed_ising — Function

convert_positive_qumo_to_mixed_ising(q::qumo{T})::mixed_ising{T} where {T<:Real}

Converts a positive QUMO problem, i.e. a problem where all variables are expected to be positive (binary in ${0,1}$, continuous in $[0,1]$) to mixed Ising, where binaries are ${-1, 1}$ and continuous in $[-1, 1]$.

The transformation is achieved by observing that we need to scale all variables $x$ by $2x-1$ to move from $[0,1]$ to $[-1,1]$. Observe, that the optimization objective is $(1/2)x'Qx + f'x + c$.

AOCoptimizer.QUMO.convert_to_qumo — Function

convert_to_qumo(q::mixed_ising{T})::qumo{T} where {T<:Real}

Converts a mixed Ising problem, i.e., a problem where binaries are ${-1, 1}$ and continuous in $[-1, 1]$ to QUMO, where binaries are in ${0, 1}$ and continuous stay at $[-1, 1]$.

AOCoptimizer.QUMO.convert_to_mixed_ising — Function

convert_to_mixed_ising(q::qumo{T})::mixed_ising{T} where {T<:Real}

Converts a QUMO problem to a mixed Ising problem. The transformation is achieved by observing that we need to map all binary variables from {0, 1} to {-1, 1}.