Reference

AOCoptimizer.AOCoptimizer — Module

AOCoptimizer is a Julia package for solving QUBO (Quadratic Unconstrained Binary Optimization) and QUMO (Quadratic Unconstrained Binary Optimization) problems.

AOCoptimizer.RuntimeUtils.run_for — Function

run_for(fn, timeout; threads)

Executes function fn for timeout seconds using threads parallel threads. The function fn should accept just one argument, which is a cancellation token; it can return any value. Returns the output of all executions.

AOCoptimizer.QUMO.mixed_ising — Type

mixed_ising{T}

A structure representing a Mixed Ising problem. Such problems are always minimization problems and, unlike traditional Ising problems, they can contain both binary and continuous variables. The continuous variables have a range of [-1, 1], while the binary variables are in {-1, 1}.

The mixed_ising structure contains:

Quadratic: A matrix representing the quadratic terms of the Ising problem.

Field: A vector representing the linear terms of the Ising problem.

Binary: A vector of booleans indicating whether each variable is binary.

Offset: A constant offset added to the objective function.

AOCoptimizer.QUBO._random — Function

_random(rng::AbstractRNG, q::qubo{T})::AbstractVector{T} where {T<:Real}

Generates a random assignment for the QUBO problem q using the provided random number generator rng.

AOCoptimizer.RuntimeUtils.RuntimeException — Type

RuntimeException

Exception type for runtime errors.

AOCoptimizer.QUBO.increase! — Function

increase!(assignment, terms, indices; ctx)

Tries to increase the objective function of an assignment applied to a set of QUBO terms by flipping assignments one-by-one according to the order in indices.

If cancellation token ctx is present, then process will continue until the entire list of indices gets visited, or cancellation token ctx gets triggered.

Method returns true if progress has been made; false otherwise

AOCoptimizer.QUMO.qumo — Type

qumo{T}

A structure representing a Quadratic Unconstrained Mixed Optimization (QUMO). Such problems can be either minimization or maximization problems. They can contain both binary and continuous variables, but the continuous variables are expected to be in the range [-1, 1] and the binary variables are in {0, 1}. The objective function is (1/2)x'Qx + f'x + c, where x is the vector of variables, Q is the quadratic matrix, f is the linear vector (if it exists), and c is a constant offset.

The qumo structure contains the following fields:

Sense: The direction of optimization, either MINIMIZATION or MAXIMIZATION.

Quadratic: A matrix representing the quadratic terms of the QUMO problem.

Field: A vector representing the linear terms of the QUMO problem.

Binary: A vector of booleans indicating whether each variable is binary.

Offset: A constant offset added to the objective function.

It is expected that the Quadratic matrix is symmetric and that the diagonal elements that correspond to binary variables are zero. (This is not checked in the constructor.)

AOCoptimizer.Direction — Type

Direction

An enum of possible values for the direction of optimization. The optimizer can either minimize (MINIMIZATION) or maximize (MAXIMIZATION) the objective function.

AOCoptimizer.QUBO.greedy_random — Function

greedy_random(rng::Union{AbstractRNG, Integer}, q::qubo{T}, ctx::Union{Nothing,CancellationToken})

Performs a greedy random search on the QUBO problem q using the provided random number generator rng. The search continues until the cancellation token ctx is triggered. Returns a tuple containing the best energy found and the corresponding assignment.

Base.isapprox — Function

isapprox(q1::qumo, q2::qumo; kwargs...)

Checks if two qumo objects are approximately equal, i.e., have the same optimization direction and approximately the same weights in the quadratic and linear terms. They should also have the same binary and continuous variables.

AOCoptimizer.create_cancellation_token — Function

create_cancellation_token()

Create a cancellation token.

AOCoptimizer.QUBO.evaluate — Function

evaluate(q::qubo, assignment)

Evaluates the objective function of the QUBO problem q for a given assignment. Returns the value of the objective function.

AOCoptimizer.QUBO.size — Function

size(q::qubo)::Integer

Returns the size of the QUBO problem represented by q.

AOCoptimizer.QUBO.qubo — Type

qubo{T}

A structure representing a Quadratic Unconstrained Binary Optimization (QUBO) problem. The qubo structure contains the following fields:

Sense: The direction of optimization, either MINIMIZATION or MAXIMIZATION.

Terms: A matrix representing the quadratic terms of the QUBO problem.

AOCoptimizer.cancel! — Function

cancel!(ctx)

Send a cancellation message using cancellation token ctx.

AOCoptimizer.QUMO.number_of_variables — Function

number_of_variables(q::Union{qumo{T}, mixed_ising{T}})::Integer where {T<:Real}

Returns the number of variables in the QUMO or Mixed Ising problem q.

AOCoptimizer.is_cancelled — Function

is_cancelled(ctx)

Checks whether cancellation token ctx has been cancelled

AOCoptimizer.QUMO.convert_mixed_ising_to_positive_qumo — Function

convert_mixed_ising_to_positive_qumo(m::mixed_ising{T})::qumo{T} where T<:Real

Converts a mixed Ising problem, i.e. a problem where binaries are ${-1, 1}$ and continuous in $[-1, 1]$ to positive QUMO, where all variables are expected to be positive (binary in ${0,1}$, continuous in $[0,1]$).

The transformation is achieved by observing that we need to scale all variables $x$ by $(x+1)/2$ to move from $[-1,1]$ to $[0,1]$. Observe, that the optimization objective is $(1/2)x'Qx + f'x + c$.

AOCoptimizer.QUMO.convert_positive_qumo_to_mixed_ising — Function

convert_positive_qumo_to_mixed_ising(q::qumo{T})::mixed_ising{T} where {T<:Real}

Converts a positive QUMO problem, i.e. a problem where all variables are expected to be positive (binary in ${0,1}$, continuous in $[0,1]$) to mixed Ising, where binaries are ${-1, 1}$ and continuous in $[-1, 1]$.

The transformation is achieved by observing that we need to scale all variables $x$ by $2x-1$ to move from $[0,1]$ to $[-1,1]$. Observe, that the optimization objective is $(1/2)x'Qx + f'x + c$.

AOCoptimizer.CancellationToken — Type

Cancellation tokens can be used to enable the caller of a task to cancel the task execution

AOCoptimizer.QUMO.convert_to_qumo — Function

convert_to_qumo(q::mixed_ising{T})::qumo{T} where {T<:Real}

Converts a mixed Ising problem, i.e., a problem where binaries are ${-1, 1}$ and continuous in $[-1, 1]$ to QUMO, where binaries are in ${0, 1}$ and continuous stay at $[-1, 1]$.

AOCoptimizer.QUBO.decrease! — Function

decrease(assignment, terms, indices; ctx)

Tries to decrease the objective function of an assignment applied to a set of QUBO terms by flipping assignments one-by-one according to the order in indices.

If cancellation token ctx is present, then process will continue until the entire list of indices gets visited, or cancellation token ctx gets triggered.

Method returns true if progress has been made; false otherwise

AOCoptimizer.QUMO.convert_to_mixed_ising — Function

convert_to_mixed_ising(q::qumo{T})::mixed_ising{T} where {T<:Real}

Converts a QUMO problem to a mixed Ising problem. The transformation is achieved by observing that we need to map all binary variables from {0, 1} to {-1, 1}.